refactor(ui): split ConfigPicker into types + controls + list
ConfigPicker.tsx was 741 lines — well over the project's red-flag
size budget — and mixed four concerns: the main hook-driven component,
the persisted-pref + ConfigItem union type, the source-filter chip +
view-density toggle, and the card/row presenters for the unified list.
Splitting them gives each file a single concept and keeps the next
PR that touches the picker focused.

Layout:
- ConfigPicker.tsx (377)         — main component
- ConfigPicker.types.ts (116)    — ConfigItem union, ConfigPickerProps,
                                   ViewMode / SourceFilter, type->icon
                                   + type->tint tables, readPref /
                                   writePref localStorage helpers
- ConfigPickerControls.tsx (59)  — SourceChip + ViewToggle
- ConfigPickerList.tsx (250)     — ConfigsList + ConfigCard + ConfigRow
                                   + the sourceTagFor presenter helper

Pure code move; behaviour, props, and DOM shape are unchanged. biome /
tsc / vitest (65 tests) all pass; full make build produces a working
bundle.
